// FrontEnd/src/app/model/memberParameter.ts
/**
 * Walk To Fly API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * 建立會員參數
 */
export interface MemberParameter { 
    /**
     * 會員Id
     */
    memberId?: string;
    /**
     * 姓氏
     */
    firstName?: string;
    /**
     * 名稱
     */
    lastName?: string;
    /**
     * 密碼
     */
    passWord?: string;
    /**
     * 信箱
     */
    email?: string;
    /**
     * 生日
     */
    birthDay?: Date;
    /**
     * 性別
     */
    sex?: boolean;
    /**
     * 手機
     */
    mobilePhone?: string;
    /**
     * 市話
     */
    telePhone?: string;
    /**
     * 縣市
     */
    city?: number;
    /**
     * 鄉鎮市區
     */
    area?: number;
    /**
     * 地址
     */
    address?: string;
}